+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased
 
 ### Changed
 * Changed focus behaviour of DropdownButton and SplitButton such that focus stays on button when dropdown is opened via mouse click.
+* Update docs to use mdx syntax.
 
 ### Fixed
 * Fixed primary button issue of not getting focused on click in Firefox, Safari and IE.