+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+1.10.0 - (February 4, 2020)
+------------------
 ### Changed
 * Update docs to use mdx syntax.